backend/pyproject.toml
[project]
name = "smart_virtual_wardrobe"
version = "0.1.0"
description = ""
authors = [
    {name = "Preveen S",email = "spreveen123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "motor>=3.3.0",
    "pydantic[dotenv]>=2.0.0",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-genai (>=1.11.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pymongo>=4.0.0",
    "email-validator>=2.0.0",
    "cloudinary==1.32.0"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

